% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataClean.R
\name{DataClean}
\alias{DataClean}
\title{Remove the NA values from a data set}
\usage{
DataClean(data_set)
}
\arguments{
\item{data_set}{A data set inputted by the user. Presumably a data set that contains NA values.}
}
\value{
clean This is the data set with removed NA values. Returns a statement if there are no NA values.
}
\description{
Uses a single data set input, then uses the "na.omit" function to remove NA values.
Returns the data set with removed NA values. Returns a statement if there are no NA values.
}
