% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForFunFunction.R
\name{FunFunction}
\alias{FunFunction}
\title{Find out if an input is a number or not}
\usage{
FunFunction(number)
}
\arguments{
\item{number}{Any input by the user. Does not have to be a numeric value.}
}
\value{
Either "THIS IS INDEED A NUMBER" or "THIS IS NOT A NUMBER. TRY AGAIN." These are the only two returns.
}
\description{
Uses a single input and determines whether or not the input is a numeric value or not.
Returns the answer as a statement.
}
