% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geom_pointPlot.R
\name{graph_project_3}
\alias{graph_project_3}
\title{Generate a plot of weight versus sex}
\usage{
graph_project_3(data, sex, weight)
}
\arguments{
\item{data}{Dataset}

\item{sex}{Sex of the species}

\item{weight}{Weight of the organism}
}
\value{
gp3 Plot of weight versus sex
}
\description{
Use the surveys dataset to generate a plot of weight versus sex.
Returns a plot of weight versus sex.
}
