% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveNAs.R
\name{remove_nas_p3}
\alias{remove_nas_p3}
\title{Remove NA values from a data set}
\usage{
remove_nas_p3(data, hindfoot_length)
}
\arguments{
\item{data}{Dataset}

\item{hindfoot_length}{Hind foot length of an organism}
}
\value{
rna_p3 Selected column from a data set containing no NA values
}
\description{
Remove NA values from a specific column in a data set.
Returns a data set containing no NA values.
}
