% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForFunFunction.R
\name{FunFunction}
\alias{FunFunction}
\title{Find the sum of a column if the value is greater than 6500.}
\usage{
FunFunction(data, column)
}
\arguments{
\item{data}{A data set input by the user. In this case, it would be the crabs data set.}

\item{column}{A column contained in the inputted data set. For a numeric output, the column must contain numeric values.}
}
\value{
total The sum of all values in the inputted column. If this return is not possible, a message will be displayed.
}
\description{
Uses the input of a data set and a specific column to find its sum only if it is larger than 6500.
Returns the answer a numeric value if the sum is greater than 6500. If the sum is smaller than 6500, a message is returned.
}
